requirements.txt

openai
requests
python-telegram-bot==13.7
gspread
oauth2client
beautifulsoup4
